requirements.txt
streamlit
pinecone-client
langchain
langchain-pinecone
python-dotenv